+++ src/core/filters/admin/tabs.php
@@ -22,10 +22,10 @@ function tabs(array $tabs): array
         'general' => _x('General', 'Label', 'kk-star-ratings'),
         'appearance' => _x('Appearance', 'Label', 'kk-star-ratings'),
         'rich_snippets' => _x('Rich Snippets', 'Label', 'kk-star-ratings'),
-        // 'tab' => [
-        //     'name' => _x('Name', 'Label', 'kk-star-ratings'),
-        //     'is_disabled' => true,
-        //     'is_addon' => false,
-        // ],
+        'custom_stars' => [
+            'name' => 'Custom Stars',
+            'is_addon' => true,
+            'link' => add_query_arg(['page' => 'kk-star-ratings-addons'], admin_url('admin.php')),
+        ],
     ];
 }