+++ src/core/wp/actions/plugins_loaded.php
@@ -11,7 +11,6 @@
 
 namespace Bhittani\StarRating\core\wp\actions;
 
-use function Bhittani\StarRating\core\functions\action;
 use function Bhittani\StarRating\core\wp\functions\activate;
 
 if (! defined('KK_STAR_RATINGS')) {
@@ -24,7 +23,4 @@ function plugins_loaded()
     // We aren't using `register_activation_hook` because it is buggy
     // and does not get called when the plugin is implictly updated.
     activate();
-
-    // Let everyone know that the plugin is loaded and initialized.
-    action('init', kksr());
 }